refactor(tools): 分离Page系列类
